+++ app/src/main/java/labs/lucka/mlp/MainActivity.kt
@@ -1,5 +1,6 @@
 package labs.lucka.mlp
 
+import android.app.Activity
 import android.content.Intent
 import android.os.Build
 import android.os.Bundle
@@ -11,6 +12,7 @@ import android.view.Menu
 import android.view.MenuItem
 import kotlinx.android.synthetic.main.activity_main.*
 import kotlinx.android.synthetic.main.content_main.*
+import org.jetbrains.anko.defaultSharedPreferences
 
 /**
  * MainActivity for MLP
@@ -20,17 +22,28 @@ import kotlinx.android.synthetic.main.content_main.*
  * - Replace saveData and loadData with [DataKit]
  * - Migrate isServiceOnline] and isMockLocationEnabled to [MockLocationProviderService]
  * - Migrate showDeveloperOptionsDialog and showAddMockTargetDialog to [DialogKit]
+ * ### 0.2.4
+ * - Import / export JSON (Totally ready)
+ * - Import / export GPX (UI only)
  *
  * ## Private Attributes
  * - [mockTargetList]
  * - [mainRecyclerViewAdapter]
  * - [mainRecyclerViewListener]
  *
+ * ## Nested Classes
+ * - [AppRequest]
+ *
  * ## Methods
  * ### Overridden
  * - [onCreate]
+ * - [onCreateOptionsMenu]
+ * - [onOptionsItemSelected]
+ * - [onActivityResult]
  * ### Private
  * - [updateFabService]
+ *
+ * @see <a href="https://www.techotopia.com/index.php/An_Android_Storage_Access_Framework_Example#Saving_to_a_Storage_File">An Android Storage Access Framework Example | Techotopia</a>
  *
  * @author lucka-me
  * @since 0.1
@@ -64,18 +77,32 @@ class MainActivity : AppCompatActivity() {
 
         }
 
+    /**
+     * Request codes
+     *
+     * @author lucka-me
+     * @since 0.2.4
+     */
+    enum class AppRequest(val code: Int) {
+        IMPORT_GPX(3101),
+        IMPORT_JSON(3102),
+        EXPORT_GPX(3201),
+        EXPORT_JSON(3202)
+    }
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
         setSupportActionBar(toolbar)
 
-        //loadData()
+        // Load data
         try {
             mockTargetList = DataKit.loadData(this)
         } catch (error: Exception) {
             DialogKit.showSimpleAlert(this@MainActivity, error.message)
         }
 
+        // Setup recycler view
         mainRecyclerViewAdapter =
             MainRecyclerViewAdapter(this, mockTargetList, mainRecyclerViewListener)
         mainRecyclerView.layoutManager = LinearLayoutManager(this)
@@ -83,6 +110,7 @@ class MainActivity : AppCompatActivity() {
         mainRecyclerView.isNestedScrollingEnabled = false
         mainRecyclerViewAdapter.attachItemTouchHelperTo(mainRecyclerView)
 
+        // Setup fabs
         fabAddMockTarget.setOnClickListener { _ ->
             DialogKit.showAddMockTargetDialog(this, mockTargetList) {
                 mainRecyclerViewAdapter.notifyAddMockTarget()
@@ -95,7 +123,7 @@ class MainActivity : AppCompatActivity() {
         }
 
         nestedScrollView.setOnScrollChangeListener(
-            NestedScrollView.OnScrollChangeListener { _, scrollX, scrollY, oldScrollX, oldScrollY ->
+            NestedScrollView.OnScrollChangeListener { _, _, scrollY, _, oldScrollY ->
                 if (scrollY > oldScrollY) {
                     fabAddMockTarget.hide()
                 } else {
@@ -147,6 +175,26 @@ class MainActivity : AppCompatActivity() {
         // as you specify a parent activity in AndroidManifest.xml.
 
         when(item.itemId) {
+            R.id.menu_main_import -> {
+                DialogKit.showImportExportMenu(this, R.string.import_title) { fileType ->
+                    startActivityForResult(
+                        Intent(Intent.ACTION_GET_CONTENT)
+                            .addCategory(Intent.CATEGORY_OPENABLE)
+                            .setType(getString(fileType.mime)),
+                        fileType.importRequestCode
+                    )
+                }
+            }
+            R.id.menu_main_export -> {
+                DialogKit.showImportExportMenu(this, R.string.export_title) { fileType ->
+                    startActivityForResult(
+                        Intent(Intent.ACTION_CREATE_DOCUMENT)
+                            .addCategory(Intent.CATEGORY_OPENABLE)
+                            .setType(getString(fileType.mime)),
+                        fileType.exportRequestCode
+                    )
+                }
+            }
             R.id.menu_main_preference -> {
                 startActivity(Intent(this, PreferenceMainActivity::class.java))
             }
@@ -159,6 +207,80 @@ class MainActivity : AppCompatActivity() {
             R.id.menu_main_preference -> true
             else -> super.onOptionsItemSelected(item)
         }
+    }
+
+    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+        when (requestCode) {
+
+            AppRequest.IMPORT_GPX.code,
+            AppRequest.IMPORT_JSON.code -> {
+                if (resultCode == Activity.RESULT_OK && data != null) {
+                    try {
+                        val newList = when (requestCode) {
+                            AppRequest.IMPORT_GPX.code -> {
+                                DataKit.importFromGPX(DataKit.readFile(this, data.data))
+                            }
+                            AppRequest.IMPORT_JSON.code -> {
+                                DataKit.importFromJSON(DataKit.readFile(this, data.data))
+                            }
+                            else -> return
+                        }
+                        if (newList.isNotEmpty()) {
+                            mockTargetList.addAll(newList)
+                            mainRecyclerViewAdapter.notifyItemRangeInserted(
+                                mockTargetList.size - 1 - newList.size, newList.size
+                            )
+                            DataKit.saveData(this, mockTargetList)
+                            Snackbar.make(
+                                nestedScrollView,
+                                String.format(getString(R.string.import_success), newList.size),
+                                Snackbar.LENGTH_LONG
+                            ).show()
+                        } else {
+                            Snackbar.make(
+                                nestedScrollView, R.string.import_empty, Snackbar.LENGTH_LONG
+                            ).show()
+                        }
+                    } catch (error: Exception) {
+                        DialogKit.showSimpleAlert(
+                            this,
+                            String.format(getString(R.string.import_failed), error.message)
+                        )
+                    }
+                }
+            }
+
+            AppRequest.EXPORT_GPX.code,
+            AppRequest.EXPORT_JSON.code -> {
+                if (resultCode == Activity.RESULT_OK && data != null) {
+                    var exportTargetList: ArrayList<MockTarget> = ArrayList(0)
+                    if (defaultSharedPreferences.getBoolean(
+                            getString(R.string.pref_ie_export_enabled_only_key), false
+                        )) {
+                        exportTargetList = mockTargetList
+                    } else {
+                        for (target in mockTargetList)
+                            if (target.enabled) exportTargetList.add(target)
+                    }
+                    try {
+                        DataKit.writeFile(
+                            this,
+                            when (requestCode) {
+                                AppRequest.EXPORT_GPX.code ->
+                                    DataKit.exportToGPX(exportTargetList)
+                                AppRequest.EXPORT_JSON.code ->
+                                    DataKit.exportToJSON(exportTargetList)
+                                else -> return
+                            },
+                            data.data
+                        )
+                    } catch (error: Exception) {
+                        DialogKit.showSimpleAlert(this, error.message)
+                    }
+                }
+            }
+        }
+        super.onActivityResult(requestCode, resultCode, data)
     }
 
     /**